image-service/requirements.txt
fastapi[standard]==0.115.8
pydantic==2.10.6
pillow==11.2.1
